fixes for goreleaser for windows
